+++ rest_framework_services/types/__init__.py
@@ -8,6 +8,7 @@ inputs/outputs are not coupled to the mutation helpers themselves.
 from rest_framework_services.types.change_result import ChangeResult
 from rest_framework_services.types.child_collection_change import ChildCollectionChange
 from rest_framework_services.types.child_spec import ChildSpec
+from rest_framework_services.types.dispatch_result import DispatchResult
 from rest_framework_services.types.field_change import FieldChange
 from rest_framework_services.types.http_extras import HttpExtras
 from rest_framework_services.types.no_input import NoInput
@@ -22,6 +23,7 @@ __all__ = [
     "ChangeResult",
     "ChildCollectionChange",
     "ChildSpec",
+    "DispatchResult",
     "FieldChange",
     "HttpExtras",
     "NoInput",